old-tests/submission/W3C/domhtml/HTMLTableRowElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLEROW</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-tr-element.html#dom-tr-sectionrowindex" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<table align="center" summary="Table 1">
<tbody><tr>
<th>Id</th>
<th>Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
</tbody></table>
<table align="center" bgcolor="#ff0000" border="4" cellpadding="2" cellspacing="2" frame="border" id="table-1" rules="all" summary="HTML Control Table" width="680">
<caption align="top">Table Caption</caption>
<thead align="center" valign="middle">
<tr align="center" bgcolor="#00FFFF" char="*" charoff="1" valign="middle">
<th id="header-1">Employee Id</th>
<th abbr="maiden" align="center" axis="center" bgcolor="#00FFFF" colspan="1" headers="header-1" height="50" id="header-2" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="100">Employee Name</th>
<th>Position</th>
<th>Salary</th>
<th>Gender</th>
<th>Address</th>
</tr>
</thead>
<tfoot align="center" valign="middle">
<tr>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
</tr>
</tfoot>
<tbody align="center" valign="middle">
<tr>
<td abbr="maiden2" align="center" axis="center" bgcolor="#FF0000" colspan="1" headers="header-2" height="50" id="Table-3" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="175">EMP0001</td>
<td headers="header-2">Margaret Martin</td>
<td>Accountant</td>
<td>56,000</td>
<td>Female</td>
<td>1230 North Ave. Dallas, Texas 98551</td>
</tr>
<tr>
<td>EMP0002</td>
<td>Martha Raynolds</td>
<td>Secretary</td>
<td>35,000</td>
<td>Female</td>
<td>1900 Dallas Road Dallas, Texas 98554</td>
</tr>
</tbody>
</table>



<script>
var domtest = async_test("The sectionRowIndex attribute specifies the index of this row, relative        to the current section(THEAD, TFOOT, or TBODY),starting from 0.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vsectionrowindex;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("tr");
      assert_equals(nodeList.length,5);
testNode = nodeList.item(2);
      vsectionrowindex = testNode.sectionRowIndex;

      assert_equals(vsectionrowindex,0);
       
    domtest.done();
}));

</script>
<p>The sectionRowIndex attribute specifies the index of this row, relative   
    to the current section(THEAD, TFOOT, or TBODY),starting from 0. 

    Retrieve the third TR(1st In TFOOT) element within the document and 
    examine its sectionRowIndex value.</p>
<div id="log"></div>
</body></html>
